Add collision to interior
Separated interior objects into two maps.
One for objects with collision, and one without.

Currently uses a Tilemap Collider 2D and a Composite Collider 2D. Some hitboxes are not perfect, but I'll adjust the tilemap collision thingies on a separate branch

Remove Edge collider from walls

Isn't that nice of a way to handle it