Random dirt placement (#10)
* Random dirt placement

Random sprites are loaded from "Resources/DirtSpots" at startup of the game.

When a scene is loaded, the "DirtSpotDistributor" prefab places them at the same position they were before. If it is the first time that the scene is loaded, and the dirt spots were not placed yet, they are randomly placed.

* Add basic CleaningTask script

It is triggered when a player is near an object to clean.